# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RESIT_Immo API",
        "description": "API REST pour la gestion locative — propriétaires, biens, locataires et contrats.",
        "contact": {
            "name": "Support RESIT_Immo",
            "email": "support@resit-immo.fr"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
...
        "/api/proprietaires": {
            "get": {
                "tags": [
                    "Propriétaires"
                ],
                "summary": "Lister tous les propriétaires",
                "operationId": "310fba4acb97cdac5e109e5a7af7c911",
                "responses": {
                    "200": {
                        "description": "Liste des propriétaires avec leurs biens",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Proprietaire"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Propriétaires"
                ],
                "summary": "Créer un propriétaire",
                "operationId": "23c23f606ad5469f320e3319001814c2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Martin"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Sophie"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "sophie@example.com",
                                        "nullable": true
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "06 12 34 56 78",
                                        "nullable": true
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "12 Rue de Paris",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Propriétaire créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Propriétaire créé avec succès."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Proprietaire"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/proprietaires/{id}": {
            "get": {
                "tags": [
                    "Propriétaires"
                ],
                "summary": "Afficher un propriétaire avec ses biens et contrats",
                "operationId": "b31ad3867eb63592661bb3641b1fba01",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du propriétaire",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail du propriétaire",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Proprietaire"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Propriétaire introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Propriétaires"
                ],
                "summary": "Mettre à jour un propriétaire (remplacement complet)",
                "operationId": "0abfd9e05d3260c65a5cc34be589bc77",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du propriétaire",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Martin"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Sophie"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "nullable": true
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Propriétaire mis à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Propriétaire mis à jour avec succès."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Proprietaire"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Propriétaire introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Propriétaires"
                ],
                "summary": "Supprimer un propriétaire",
                "operationId": "f5d31175712731f7113e0c48380e4228",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du propriétaire",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Propriétaire supprimé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Propriétaire supprimé avec succès."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Propriétaire introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Propriétaires"
                ],
                "summary": "Mettre à jour un propriétaire (mise à jour partielle)",
                "operationId": "8d214e14c7389d99f904fca041f3a0d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du propriétaire",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Martin"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Sophie"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Propriétaire mis à jour"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Propriétaire introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Proprietaire": {
                "title": "Propriétaire",
                "required": [
                    "nom",
                    "prenom"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "Martin"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "Sophie"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "sophie@example.com",
                        "nullable": true
                    },
                    "telephone": {
                        "type": "string",
                        "example": "06 12 34 56 78",
                        "nullable": true
                    },
                    "adresse": {
                        "type": "string",
                        "example": "12 Rue de Paris",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "title": "Erreur de validation",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "nom": [
                                "Le nom est obligatoire."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez votre token Sanctum : Bearer {token}"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ],
...
}